Refactor command registry around catalogs
